RENAMESCRIPT.SH INTEGRATED; WORKS

+++ Rename/RenameScript.sh
@@ -1,3 +1,10 @@
+#####################################################################
+#                                                                   #
+# It's a script from spschoen.  Run it like any other.              #
+# sh RenameScript.sh directory_to_rename/                           #
+#                                                                   #
+#####################################################################
+
 #!/bin/bash
 
 # Use the below if needed.
@@ -5,6 +12,7 @@
 readonly LOG_FILE="$(pwd)/$(basename "$0")_$(date +%s).log"
 info()    { echo "[INFO]    $@" | tee -a "$LOG_FILE" >&2 ; }
 fatal()   { echo "[FATAL]   $@" | tee -a "$LOG_FILE" >&2 ; exit 1 ; }
+error()   { echo "[ERROR]   $@" | tee -a "$LOG_FILE" >&2 ; }
 
 #Cleaning everything up.
 clear
@@ -29,37 +37,69 @@ rm -f *.log
 #Saving argument 1.
 DIRECTORY="$1"
 
-cd $DIRECTORY
+directoryCount=$(find "$DIRECTORY"/* -maxdepth 1 -type d | wc -l)
 
-for D in *; do
-    if [ -d "${D}" ]; then
-		if [[ "${D}" == *" "* ]]; then
-			DIR_NAME="${D// /_}"
-			info "Renaming ${D} to $DIR_NAME"
-			mv "${D}" "$DIR_NAME"
-			cd "$DIR_NAME"
-		else
-			cd "${D}"
+if [[ "$directoryCount" == "0" ]]; then
+    info "Did not detect subdirectories in ""$DIRECTORY"
+else
+    info "Detected subdirectories in ""$DIRECTORY"
+fi
+
+cd "$DIRECTORY"
+
+if [[ "$directoryCount" != "0" ]]; then
+	info "--------------------------------------------------"
+	for D in *; do
+		if [ -d "${D}" ]; then
+			if [[ "${D}" == *" "* ]]; then
+				DIR_NAME="${D// /_}"
+				info "Renaming ${D} to $DIR_NAME"
+				mv "${D}" "$DIR_NAME"
+				cd "$DIR_NAME"
+			else
+				cd "${D}"
+			fi
+			info "Currently working in: $(pwd)"
+			
+			for F in *; do
+				if [[ "${F}" == *"_"* ]]; then
+					FIL_NAME="${F##*_}"
+					info "Renaming ${F} to $FIL_NAME"
+					mv "${F}" "$FIL_NAME"
+				fi
+				# info "File: ""${F}"
+				
+				FILESIZE=$(stat -c%s "${FIL_NAME}")
+				if [[ "$FILESIZE" == "76" ]] && [[ "${F}" == *".html" ]]; then
+					warning "Detected Moodle downloaded HTML comment with no comment - deleting."
+					rm -f "$FILESIZE"
+				fi
+				
+			done
+			
+			info "--------------------------------------------------"
+			cd ..
 		fi
-		info "Currently working in: $(pwd)"
-		
-		for F in *; do
-			if [[ "${F}" == *"_"* ]]; then
-				FIL_NAME="${F##*_}"
-				info "Renaming ${F} to $FIL_NAME"
-				mv "${F}" "$FIL_NAME"
+	done
+else
+	info "--------------------------------------------------"
+	for F in *; do
+		if [ -f "${F}" ] && [ ! -d "${F}" ]; then
+			STUDENT_NAME="${F%%_*}"
+			STUDENT_NAME="${STUDENT_NAME// /_}"
+			FILE_NAME="${F##*_}"
+			# info "Original Name: ""${F}"
+			# info "Student Name : ""$STUDENT_NAME"
+			# info "File Name    : ""$FILE_NAME"
+			if [ ! -d "$STUDENT_NAME" ]; then
+				info "Did not detect directory ""$STUDENT_NAME"", creating now."
+				mkdir "$STUDENT_NAME"
 			fi
-			# info "File: ""${F}"
-			
-			FILESIZE=$(stat -c%s "${F}")
-			if [[ "$FILESIZE" == "76" ]] && [[ "${F}" == *".html" ]]; then
-				warning "Detected Moodle downloaded HTML comment with no comment - deleting."
-				rm -f "$FILESIZE"
-			fi
-			
-		done
-        
+			info "Moving ""${F}"" to ""$STUDENT_NAME""/""$FILE_NAME"
+			mv "${F}" "$STUDENT_NAME"/"$FILE_NAME"
+		fi
 		info "--------------------------------------------------"
-        cd ..
-    fi
-done
+	done
+fi
+
+
